feat: Add staff deletion and grade definition update
Introduces the ability to delete staff members and update grade definitions. This enhances the administrative capabilities of the student database by allowing for better management of both personnel and academic structures. The `deletingStaff` state has been added to `App.tsx` to manage the deletion workflow. The `handleUpdateGradeDefinition` function is also implemented in `App.tsx` to facilitate changes to grade configurations.